// src/lib/types/order.ts
import { OrderStatus } from "../enums/order.enum";
import { Product } from "./product";

export interface OrderInquiry {
    page: number;
    limit: number;
    orderStatus: OrderStatus;
}

export interface OrderItem {
    _id: string
    itemQuantity: number;
    itemPrice: number;
    orderId: string;
    productId: string;
    createAt: Date;
    updateAt: Date;
}

export interface Order {
    _id: string;
    orderTotal: number;
    orderDelivery: number;
    orderStatus: OrderStatus;
    memberId: string;
    createAt: Date;
    updateAt: Date;
    /** from aggregation **/
    orderItems: OrderItem[];
    productData: Product[];
}

export interface OrderItemInput {
    itemQuantity: number;
    itemPrice: number;
    productId: string;
    orderId?: string;
}

export interface OrderUpdateInput {
    orderId: String;
    orderStatus: OrderStatus;
}